web/requirements.txt
coverage==4.4.1
pymongo==3.5.1
marshmallow==3.0.0b3
factory_boy==2.9.2
gunicorn==19.7.1
WebOb==1.7.3
